#229 Fix decoding of POPM "user-email" field. Extract BOM field from encoding settings Parse encoding as token.
