+++ src/ogg/vorbis/Vorbis.ts
@@ -1,6 +1,6 @@
 import * as Token from 'token-types';
 
-import {AttachedPictureType} from '../../id3v2/ID3v2';
+import {AttachedPictureType} from '../../id3v2/ID3v2Token';
 import {IPicture} from '../../type';
 
 export interface IVorbisPicture extends IPicture {
